Refactor: Rename inconsistent directory.
Refs: BLAISE-3651
