+++ src/components/questionnaireDetails/sections/viewCatiModeDetails.test.tsx
