+++ src/components/questionnaireDetails/sections/viewCatiModeDetails.tsx
